requirements.txt
# Core MCP dependencies
mcp>=1.9.0
fastmcp>=0.4.0

# Environment configuration
python-dotenv>=1.0.0

# FastAPI and web framework - Use flexible versions to avoid conflicts
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Redis and caching
redis>=5.0.0

# Authentication and security
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client and utilities
httpx>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.2.0
pydantic>=2.5.0

# Docker integration
docker>=6.1.0

# Task queue
celery>=5.3.0
kombu>=5.3.0

# Configuration
pyyaml>=6.0.0

# Additional utilities - Compatible with MCP requirements
anyio>=4.6